+++ src/snowflake/snowpark/files.py
@@ -6,6 +6,7 @@
 from __future__ import annotations
 
 import array
+import io
 from io import RawIOBase
 from typing import Iterable
 
@@ -27,12 +28,15 @@ class SnowflakeFile(RawIOBase):
         self,
         file_location: str,
         mode: str = "r",
+        is_owner_file: bool = False,
     ) -> None:
         super().__init__()
         # The URL/URI of the file to be opened by the SnowflakeFile object
         self._file_location: str = file_location
         # The mode of file stream
         self._mode: str = mode
+        # Whether it is intended to access owner's files
+        self._is_owner_file = is_owner_file 
 
         # The attributes supported as part of IOBase
         self.buffer = None
@@ -45,16 +49,20 @@ class SnowflakeFile(RawIOBase):
         cls,
         file_location: str,
         mode: str = "r",
+        is_owner_file: bool = False,
     ) -> SnowflakeFile:
         """
         Returns a :class:`~snowflake.snowpark.file.SnowflakeFile`.
         In UDF and Stored Procedures, the object works like a Python IOBase object and as a wrapper for an IO stream of remote files. The IO Stream is to support the file operations defined in this class.
 
+	All files are accessed in the context of the UDF owner and therefore when accessing call files user should prefer setting is_owner_file to default false value as it only allows scoped URLs and therefore a caller cannot inadvertently or maliciously access the owner's files.
+
         Args:
             file_location: A string of file location. It can be a remote URL/URI.
             mode: A string used to mark the type of an IO stream.
+            is_owner_file: A boolean value, if True, the API is intended to access owner's files and all url/uri are allowed. If False, the API is intended to access files passed into the function by the caller and only scope url is allowed.
         """
-        return cls(file_location, mode)
+        return cls(file_location, mode, is_owner_file)
 
     def close(self) -> None:
         """
@@ -96,6 +104,9 @@ class SnowflakeFile(RawIOBase):
         raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
     def readlines(self, hint: int = -1) -> None:
+        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+
+    def seek(self, offset: int, whence: int = io.SEEK_SET) -> int:
         raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
     def seekable(self) -> None: