+++ tests/integ/scala/test_dataframe_join_suite.py
@@ -1115,3 +1115,30 @@ def test_select_columns_on_join_result_with_conflict_name(session):
     assert df4.schema.fields[1].name == "D"
     assert len(re.search('"l_.*_A"', df4.schema.fields[2].name).group(0)) > 0
     assert df4.collect() == [Row(3, 4, 1)]
+
+
+def test_join_diamond_shape_error(session):
+    """This is supposed to work but currently we don't handle it correctly. We should fix this with a good design."""
+    df1 = session.create_dataframe([[1]], schema=["a"])
+    df2 = session.create_dataframe([[1]], schema=["a"])
+    df3 = df1.join(df2, df1["a"] == df2["a"])
+    df4 = df3.select(df1["a"].as_("a"))
+    # df1["a"] and df4["a"] has the same expr_id in map expr_to_alias. When they join, only one will be in df5's alias
+    # map. It leaves the other one resolved to "a" instead of the alias.
+    df5 = df1.join(df4, df1["a"] == df4["a"])
+    with pytest.raises(
+        SnowparkSQLAmbiguousJoinException,
+        match="The reference to the column 'A' is ambiguous.",
+    ):
+        df5.collect()
+
+
+def test_join_diamond_shape_workaround(session):
+    df1 = session.create_dataframe([[1]], schema=["a"])
+    df2 = session.create_dataframe([[1]], schema=["a"])
+    df3 = df1.join(df2, df1["a"] == df2["a"])
+    df4 = df3.select(df1["a"].as_("a"))
+    # df1_converted["a"]  has a different expr_id from df4["a"]. So the join works.
+    df1_converted = df1.select(df1["a"])
+    df5 = df1_converted.join(df4, df1_converted["a"] == df4["a"])
+    Utils.check_answer(df5, [Row(1, 1)])