+++ tests/integ/test_simplifier_suite.py
@@ -27,7 +27,7 @@ from tests.utils import TestData, Utils
 
 @pytest.fixture(scope="module", autouse=True)
 def skip(pytestconfig):
-    if not pytestconfig.getoption("use_sql_simplifier"):
+    if pytestconfig.getoption("disable_sql_simplifier"):
         pytest.skip(
             "Disable sql simplifier test when simplifier is disabled",
             allow_module_level=True,
@@ -963,3 +963,23 @@ def test_natural_join(session, simplifier_table):
     df = df.select("a").select("a").select("a")
     df.collect()
     assert df.queries["queries"][0].count('SELECT "A"') == 1
+
+
+def test_rename_to_dropped_column_name(session):
+    session.sql_simplifier_enabled = True
+    df1 = session.create_dataframe([[1, 2, 3]], schema=["a", "b", "c"])
+    df2 = df1.drop("a").drop("b")
+    df3 = df2.withColumn("a", df2["c"])
+    df4 = df3.withColumn("b", sql_expr("1"))
+    assert df4.columns == ["C", "A", "B"]
+    Utils.check_answer(df4, [Row(3, 3, 1)])
+
+
+def test_rename_to_existing_column_column(session):
+    session.sql_simplifier_enabled = True
+    df1 = session.create_dataframe([[1, 2, 3]], schema=["a", "b", "c"])
+    # df2 = df1.drop("a").drop("b")
+    df3 = df1.withColumn("a", df1["c"])
+    df4 = df3.withColumn("b", sql_expr("1"))
+    assert df4.columns == ["C", "A", "B"]
+    Utils.check_answer(df4, [Row(3, 3, 1)])